Extract verifyUser to a middleware file
